requirements.txt
### Requirements
librosa
torch==1.13.1
torchaudio==0.13.1
soundfile
scipy
matplotlib
tqdm
numpy
pandas
speechbrain
tensorflow
tensorflow-probability
torchmetrics==0.10.0
seaborn
ipykernel
scaper
thop==0.1.1.post2209072238
openl3
youtube_dl
transformers
bs4
pyroomacoustics
python-sofa==0.2.0
onnx
onnxruntime
torch_tb_profiler
ffmpegio
noisereduce
